+++ changelog.md
@@ -1,6 +1,14 @@
 RestyGWT
 ========
 
+[resty-gwt 1.1]
+---------------
+
+* add ability to take part of the compilation process by registering 
+  ``org.fusesource.restygwt.rebind.AnnotationResolver`` in 
+  ``org.fusesource.restygwt.rebind.BindingDefaults``
+* add some example for triggering ModelChangeEvent from a MethodCallback
+
 [resty-gwt 1.0](http://restygwt.fusesource.org/blog/releases/2010/09/release-1-0.html), released 2010-09-27
 ----------------------------------